+++ src/include/sll/api/memory.h
@@ -1,6 +1,5 @@
 #ifndef __SLL_API_MEMORY_H__
 #define __SLL_API_MEMORY_H__ 1
-#include <sll/api.h>
 #include <sll/common.h>
 #include <sll/types.h>