+++ src/include/sll/static_object.h
@@ -5,12 +5,7 @@
 
 
 
-#define __SLL_STATIC_NAME(nm) sll_static_##nm
-#define __SLL_STATIC_OBJECT_DECL(nm) extern sll_runtime_object_t* __SLL_STATIC_NAME(nm)
-
-
-
-#define SLL_ACQUIRE_STATIC(nm) (SLL_ACQUIRE(__SLL_STATIC_NAME(nm)),__SLL_STATIC_NAME(nm))
+#define SLL_ACQUIRE_STATIC(nm) (SLL_ACQUIRE(sll_static_##nm),sll_static_##nm)
 #define SLL_ACQUIRE_STATIC_INT(v) (SLL_ACQUIRE(((v)<0?sll_static_negative_int[-(v)-1]:sll_static_int[(v)])),((v)<0?sll_static_negative_int[-(v)-1]:sll_static_int[(v)]))
 #define SLL_ACQUIRE_STATIC_CHAR(c) SLL_FROM_CHAR(c)
 #define SLL_FROM_INT(v) __SLL_ADD_DEBUG_DATA(sll_int_to_object(v),__SLL_DEBUG_TYPE_CREATE)
@@ -27,15 +22,15 @@ extern sll_runtime_object_t* sll_static_negative_int[16];
 
 
 
-__SLL_STATIC_OBJECT_DECL(float_zero);
+extern sll_runtime_object_t* sll_static_float_zero;
 
 
 
-__SLL_STATIC_OBJECT_DECL(float_one);
+extern sll_runtime_object_t* sll_static_float_one;
 
 
 
-__SLL_STATIC_OBJECT_DECL(float_half);
+extern sll_runtime_object_t* sll_static_float_half;
 
 
 
@@ -43,19 +38,19 @@ extern sll_runtime_object_t* sll_static_char[256];
 
 
 
-__SLL_STATIC_OBJECT_DECL(str_zero);
+extern sll_runtime_object_t* sll_static_str_zero;
 
 
 
-__SLL_STATIC_OBJECT_DECL(array_zero);
+extern sll_runtime_object_t* sll_static_array_zero;
 
 
 
-__SLL_STATIC_OBJECT_DECL(handle_zero);
+extern sll_runtime_object_t* sll_static_handle_zero;
 
 
 
-__SLL_STATIC_OBJECT_DECL(map_zero);
+extern sll_runtime_object_t* sll_static_map_zero;